Crate aoc_2017_day_17

Advent of Code 2017: Day 17: Spinlock

https://adventofcode.com/2017/day/17

Input

A single line containing a positive number. This is the number of positions to advance in each iteration.

Part one

Begin with a circular buffer containing only the number 0. Iterate 2017 times, moving forward the number of positions specified in the input, plus one, then insert the next number after the current position. The answer is the number after that in the resulting buffer.

Part two

Same as part one, but iterate fifty million times, and return the number following 0 in the buffer.

Algorithm

For part one, we use a naïve algorithm, creating the circular buffer and inserting the number each time.

For part two, we instead only track the number that was last inserted after zero. We model the circular buffer with only the following information:

size: total number of items in the buffer

pos: current position, where 0 is at position 0

second: item at position 1, ie the answer

Instead of iterating fifty million times, we alternate between advancing towards the end of the buffer quickly using division, and advancing past the end of the buffer in a single step. This allows the process to complete much more quickly.

Modules

input

logic

Logic for solving the puzzles

Functions

puzzle_metadata